Initial working code of sending and receiving to MailStation
This uses a new LPT interface to the MailStation that uses all 8
data lines to simplify reading and writing data, rather than using
the "tribble" style of transfer in the stock firmware.

This uses an MCP23017 in i2c mode for all of the GPIO, since there
aren't enough pins on the ESP8266.