ms_write: Go back to writing each pin individually
Sending more than a few K of data frequently gets corrupted between
the ESP8266 and the MailStation (all of the data is echoed between
ESP8266 and host fine).  Rather than writing all of the data pins in
one shot, go back to writing each pin separately.  This is obviously
slower but it seems to be more reliable.