HTTP: Add HTTP server with file upload interface
Disabled by default, can be enabled with AT$HTTP=1.

Since we have to send the file size to WSLoader before sending any
data and the file upload is streamed to us without the known file
size, use an intermediate endpoint to measure the file.

The browser will stream the file to /upload and we'll tally up the
bytes without sending anything to the MailStation, then we send a
307 redirect to the client to /upload_measured?size=xxx.  The
browser will follow the redirect and POST the file again, so we'll
have the size at the start and we can send it to WSLoader first.

This allows files to be uploaded with curl, though maybe text or
json output would be nicer.

    curl -L -F file=@msterm.bin http://192.168.2.55/upload