Telnet: If MailStation input starts with \e, buffer for a bit
This is probably a multi-character key input like an arrow key, so
buffer it and do telnet.print() with all of the input at once so
hopefully the remote end gets it all at once.

Fixes sending arrow keys and Page Up/Down to remote systems that
don't wait very long to read such input.