AT$UPLOAD: Add a hack to send a trailing null byte
For some reason when WSLoader receives the final byte it's expecting
and then does its 'jp 0x8000', the ack line doesn't go low and
ms_write() has to timeout waiting for that, which looks like a
failed upload to sendload.

This could be worked around by ms_write ignoring the ack on the
final byte, but every byte counts so if it properly fails to ack, it
should be an error.

Until I can figure out why this happens (without adding a delay
before WSLoader's final jump), send WSLoader one extra null byte so
we can ignore the result of that ms_write() and still see the proper
response to the real last byte of what sendload sent.