esp8266: AT Handling: Add ATQ and ATE, do proper outputting based on those
With echo enabled, only print \r after accepting a command, and once
the command wants to start output, it should print the \n.  With ATQ
set (quiet mode), that \n is not printed since the OK or ERROR is
not printed.

Also loop through all command input so chained command sequences
like "ATEQ1" will be parsed as "ATE0" and "ATQ1".

From WiFiPPP commit b7a04199813fe7dc6a701fe272286300f86641ab

This is not really needed on WiFiStation but it helps keep code in
sync.